api 작성
파일 삭제, 파일 추가 등록, 다운로드 sample

+++ src/main/java/com/sj/board/controller/CommentController.java
@@ -8,19 +8,19 @@ import org.springframework.web.bind.annotation.*;
 public class CommentController {
 
     @PostMapping("/posts/{postsId}/comm")
-    public String insertComment(@PathVariable int postsId, @RequestBody CommentDto comment) {
+    public String insertComment(@PathVariable long postsId, @RequestBody CommentDto comment) {
 
         return "댓글 등록";
     }
 
     @PutMapping("/posts/{postsId}/comm/{commId}")
-    public String updateComment(@PathVariable int postsid, @PathVariable int commId) {
+    public String updateComment(@PathVariable long postsid, @PathVariable long commId) {
 //
         return "updatecomment";
     }
 
     @DeleteMapping("/posts/{postsId}/comm/{commId}")
-    public String deleteComment(@PathVariable int postsId, @PathVariable int commId) {
+    public String deleteComment(@PathVariable long postsId, @PathVariable long commId) {
 
         return "deleteComment";
     }
